cadling/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cadling"
version = "0.1.0"
description = "CAD document processing toolkit inspired by docling"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "LatticeLabs"},
]
keywords = [
    "CAD",
    "STEP",
    "STL",
    "document-processing",
    "optical-recognition",
    "neural-networks",
    "RAG",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # Core dependencies
    "pydantic>=2.0,<3.0",
    "numpy>=1.24,<2.0",
    "pillow>=10.0",

    # Typing
    "typing-extensions>=4.0",

    # Logging
    "rich>=13.0",  # For better logging
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-xdist>=3.0",  # Parallel testing
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "pre-commit>=3.0",
]

# CAD processing dependencies (require conda or special setup)
cad = [
    # Note: pythonocc-core must be installed via conda
    # conda install -c conda-forge pythonocc-core
    "numpy-stl>=3.0",
    "trimesh>=4.0",
    "networkx>=3.0",  # For topology graphs
]

# Neural network dependencies
ml = [
    "torch>=2.0",
    "torch-geometric>=2.3",
    "transformers>=4.35",
]

# Vision model dependencies
vision = [
    "transformers>=4.35",
    "easyocr>=1.7",
    "opencv-python>=4.8",
]

# All optional dependencies
all = [
    "cadling[dev,cad,ml,vision]",
]

[project.urls]
Homepage = "https://github.com/yourusername/cadling"
Documentation = "https://cadling.readthedocs.io"
Repository = "https://github.com/yourusername/cadling"
Issues = "https://github.com/yourusername/cadling/issues"

[project.scripts]
cadling = "cadling.cli.main:cli"

[tool.setuptools]
packages = ["cadling"]

[tool.setuptools.package-data]
cadling = ["py.typed"]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 88
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "B905",  # zip strict parameter
]
exclude = [
    ".git",
    ".venv",
    "build",
    "dist",
    "__pycache__",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "F403"]  # Allow unused imports in __init__

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start lenient, tighten later
ignore_missing_imports = true
exclude = [
    "tests/",
    "build/",
    "dist/",
]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=cadling",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "requires_gpu: marks tests that require GPU",
    "requires_pythonocc: marks tests that require pythonocc-core",
]

[tool.coverage.run]
source = ["cadling"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
